// src/channels/channels.d.ts
export type OverwriteType = {
  id: string; //	role or user id
  type: number; //	either 0 (role) or 1 (member)
  allow: string; //	permission bit set
  deny: string; //	permission bit set
};